pyproject.toml
[tool.poetry]
name = "opg-pipeline-builder"
version = "0.1.0"
description = ""
authors = ["gwionap <gwion.aprhobat@justice.gov.uk>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
python-dateutil = "^2.8.2"
boto3 = "^1.24.77"
dataengineeringutils3 = "^1.4.1"
pyarrow = ">=8.0.0, <9.0.0"
mojap-metadata = "^1.10.2"
awswrangler = "^2.17.0"
data-linter = "^6.1.1"
croniter = "^1.3.7"
pydbtools = "^5.3.1"
gluejobutils = "^3.2.0"
etl-manager = "^7.5.2"
ssm-parameter-store = "^19.11.0"
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
moto = "^3.1.6"
pytest-cov = "^3.0.0"
duckdb = "^0.5.1"
sqlglot = "^6.2.3"
pre-commit = "^2.20.0"
detect-secrets = "^1.4.0"
mkdocs-material = "^8.5.6"
mkdocstrings = "^0.19.0"
mkdocstrings-python = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
